Updated to make programs and variables fully functional.
git-svn-id: file:///p/svn/Projects/Android/ISYController@111 1e7840b9-f7c2-db11-b69c-00d0b71b9132

+++ src/com/pataelmo/isycontroller/DatabaseHelper.java
@@ -365,7 +365,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 				c.getString(c.getColumnIndex(KEY_TYPE)),
 				c.getInt(c.getColumnIndex(KEY_INIT)),
 				c.getInt(c.getColumnIndex(KEY_VALUE)),
-				c.getInt(c.getColumnIndex(KEY_LASTCHANGED)));
+				c.getLong(c.getColumnIndex(KEY_LASTCHANGED)));
 		c.close();
 		return result;
 	}
